+++ src/Stack/Types/Build.hs
@@ -426,6 +426,8 @@ data BuildOpts =
             ,boptsExeProfile :: !Bool
             ,boptsHaddock :: !Bool
             -- ^ Build haddocks?
+            ,boptsOpenHaddock :: !Bool
+            -- ^ Open local haddocks on the browser?
             ,boptsHaddockDeps :: !(Maybe Bool)
             -- ^ Build haddocks for dependencies?
             ,boptsDryrun :: !Bool
@@ -469,6 +471,7 @@ defaultBuildOpts = BuildOpts
     , boptsLibProfile = False
     , boptsExeProfile = False
     , boptsHaddock = False
+    , boptsOpenHaddock = False
     , boptsHaddockDeps = Nothing
     , boptsDryrun = False
     , boptsGhcOptions = []